requirements.txt
pytest~=7.2.0
flake8
djangorestframework